Ensure decoding of messages with content-type set
Previously, if a string could be encoded as UTF-8 in a valid manner it
was assumed that it was already decoded regardless of specified
content-type. This unfortunately meant that for encodings that happened
to be encoded in valid UTF-8 characters (like ISO-2022-JP), the message
content was never properly decoded.